Added cost calculation during validation and added execution building of actions.

+++ Source/Goap/Actions/Abstraction/GoapActionBase.cs
@@ -55,7 +55,7 @@ public abstract class GoapActionBase : IGoapAction
         return conditionToMeet;
     }
     
-    public int CalculateCost()
+    public float CalculateCost()
     {
         return ActionDataComponent.CalculatedCost;
     }
