+++ Source/Goap/Actions/ActionComponents/GoapActionDataComponent.cs
@@ -6,5 +6,5 @@ public class GoapActionDataComponent
 {
     public float TimeCostMultiplier { get; init; }
     public float TimeCostInSeconds { get; set; }
-    public int CalculatedCost => (int) (TimeCostMultiplier * TimeCostInSeconds);
+    public float CalculatedCost => TimeCostMultiplier * TimeCostInSeconds;
 }
